+++ src/main/java/com/midasdev/mochat/auth/service/AuthService.java
@@ -14,6 +14,9 @@ import com.midasdev.mochat.config.security.jwt.RefreshToken;
 import com.midasdev.mochat.config.security.jwt.TokenAttribute;
 import com.midasdev.mochat.config.security.jwt.TokenType;
 import com.midasdev.mochat.config.security.jwt.repository.RefreshTokenRedisRepository;
+import com.midasdev.mochat.global.exception.ApplicationException;
+import com.midasdev.mochat.global.exception.ApplicationExceptionType;
+import com.midasdev.mochat.global.util.assertion.Assertion;
 import com.midasdev.mochat.member.domain.Member;
 import com.midasdev.mochat.member.service.MemberService;
 import lombok.RequiredArgsConstructor;
@@ -50,6 +53,18 @@ public class AuthService {
         AuthorizationToken authorizationToken = jwtProvider.createAuthorizationToken(memberId);
         refreshTokenRedisRepository.save(RefreshToken.from(memberId, authorizationToken));
         return authorizationToken;
+    }
+
+    public void verifyRefreshToken(Long memberId, String refreshToken) {
+        // TEST: 저장된 refreshToken 여부에 대한 테스트
+        // TEST: refreshToken 일치 여부에 대한 테스트
+        // REFACTOR: AuthService로 이동
+        RefreshToken refreshTokenFromRedis = refreshTokenRedisRepository.findById(memberId)
+                                                                        .orElseThrow(() -> new ApplicationException(
+                                                                                ApplicationExceptionType.REFRESH_TOKEN_NOT_FOUND, memberId));
+        Assertion.with(refreshToken)
+                 .setValidation(refreshTokenFromRedis::isSameToken)
+                 .validateOrThrow(() -> new ApplicationException(ApplicationExceptionType.REFRESH_TOKEN_MISMATCH, memberId));
     }
 
     public void logoutMember(Long memberId) {
