+++ src/main/java/com/midasdev/mochat/member/service/MemberService.java
@@ -1,11 +1,9 @@
 package com.midasdev.mochat.member.service;
 
 import com.midasdev.mochat.auth.dto.TokenRequestUser;
-import com.midasdev.mochat.config.security.jwt.RefreshToken;
 import com.midasdev.mochat.config.security.jwt.repository.RefreshTokenRedisRepository;
 import com.midasdev.mochat.global.exception.ApplicationException;
 import com.midasdev.mochat.global.exception.ApplicationExceptionType;
-import com.midasdev.mochat.global.util.assertion.Assertion;
 import com.midasdev.mochat.member.domain.Member;
 import com.midasdev.mochat.member.repository.MemberSpringDataRepository;
 import java.util.Optional;
@@ -40,18 +38,6 @@ public class MemberService {
                               .build();
 
         return memberSpringDataRepository.save(member);
-    }
-
-    public void verifyRefreshToken(Long memberId, String refreshToken) {
-        // TEST: 저장된 refreshToken 여부에 대한 테스트
-        // TEST: refreshToken 일치 여부에 대한 테스트
-        // REFACTOR: AuthService로 이동
-        RefreshToken refreshTokenFromRedis = refreshTokenRedisRepository.findById(memberId)
-                                                                        .orElseThrow(() -> new ApplicationException(
-                                                                                ApplicationExceptionType.REFRESH_TOKEN_NOT_FOUND, memberId));
-        Assertion.with(refreshToken)
-                 .setValidation(refreshTokenFromRedis::isSameToken)
-                 .validateOrThrow(() -> new ApplicationException(ApplicationExceptionType.REFRESH_TOKEN_MISMATCH, memberId));
     }
 
 }
